fix phpstan errors on level 3

+++ src/classes/Core/Debugging.php
@@ -166,7 +166,7 @@ class Debugging {
 	 *
 	 * @since  6.0.0
 	 * @param  string $type Type of count, values: total|pages.
-	 * @return integer
+	 * @return int
 	 */
 	public function get_logs_count( $type = 'total' ) {
 		global $wpdb;
@@ -174,7 +174,7 @@ class Debugging {
 		$total = $wpdb->get_var( 'SELECT FOUND_ROWS();' ); //phpcs:ignore
 
 		if ( 'pages' === $type ) {
-			return ceil( $total / $this->logs_per_page );
+			return (int) ceil( $total / $this->logs_per_page );
 		}
 
 		return $total;
